Fix polymorphism.md#interface section
