+++ guide/example/example-poly-10.kt
@@ -22,7 +22,10 @@ interface Project {
 @SerialName("owned")
 class OwnedProject(override val name: String, val owner: String) : Project
 
+@Serializable
+class Data(val project: Project) // Project is an interface
+
 fun main() {
-    val data: Project = OwnedProject("kotlinx.coroutines", "kotlin")
-    println(format.encodeToString(PolymorphicSerializer(Project::class), data))
-}    
+    val data = Data(OwnedProject("kotlinx.coroutines", "kotlin"))
+    println(format.encodeToString(data))
+}
