+++ guide/example/example-poly-11.kt
@@ -14,18 +14,16 @@ val module = SerializersModule {
 
 val format = Json { serializersModule = module }
 
-interface Project {
-    val name: String
+@Serializable
+abstract class Project {
+    abstract val name: String
 }
-
+            
 @Serializable
 @SerialName("owned")
-class OwnedProject(override val name: String, val owner: String) : Project
-
-@Serializable
-class Data(val project: Project) // Project is an interface
+class OwnedProject(override val name: String, val owner: String) : Project()
 
 fun main() {
-    val data = Data(OwnedProject("kotlinx.coroutines", "kotlin"))
+    val data: Any = OwnedProject("kotlinx.coroutines", "kotlin")
     println(format.encodeToString(data))
-}        
+}
