+++ guide/example/example-poly-12.kt
@@ -7,11 +7,10 @@ import kotlinx.serialization.json.*
 import kotlinx.serialization.modules.*
 
 val module = SerializersModule {
-    polymorphic(Project::class) {
+    polymorphic(Any::class) {
         subclass(OwnedProject::class)
     }
 }
-
 val format = Json { serializersModule = module }
 
 @Serializable
