+++ guide/example/example-poly-13.kt
@@ -11,6 +11,7 @@ val module = SerializersModule {
         subclass(OwnedProject::class)
     }
 }
+
 val format = Json { serializersModule = module }
 
 @Serializable
@@ -24,5 +25,5 @@ class OwnedProject(override val name: String, val owner: String) : Project()
 
 fun main() {
     val data: Any = OwnedProject("kotlinx.coroutines", "kotlin")
-    println(format.encodeToString(data))
+    println(format.encodeToString(PolymorphicSerializer(Any::class), data))
 }
