+++ guide/example/example-poly-16.kt
@@ -5,34 +5,44 @@ import kotlinx.serialization.*
 import kotlinx.serialization.json.*
 
 import kotlinx.serialization.modules.*
-import kotlin.reflect.KClass
 
-val module = SerializersModule {
+@Serializable
+abstract class Response<out T>
+            
+@Serializable
+@SerialName("OkResponse")
+data class OkResponse<out T>(val data: T) : Response<T>()
+
+val responseModule = SerializersModule {
+    polymorphic(Response::class) {
+        subclass(OkResponse.serializer(PolymorphicSerializer(Any::class)))
+    }
+}
+
+val projectModule = SerializersModule {
     fun PolymorphicModuleBuilder<Project>.registerProjectSubclasses() {
         subclass(OwnedProject::class)
     }
     polymorphic(Any::class) { registerProjectSubclasses() }
     polymorphic(Project::class) { registerProjectSubclasses() }
-}        
-
-val format = Json { serializersModule = module }
-
-interface Project {
-    val name: String
 }
 
 @Serializable
-@SerialName("owned")
-class OwnedProject(override val name: String, val owner: String) : Project
-
+abstract class Project {
+    abstract val name: String
+}
+            
 @Serializable
-class Data(
-    val project: Project,
-    @Polymorphic val any: Any 
-)
+@SerialName("OwnedProject")
+data class OwnedProject(override val name: String, val owner: String) : Project()
+
+val format = Json { serializersModule = projectModule + responseModule }
 
 fun main() {
-    val project = OwnedProject("kotlinx.coroutines", "kotlin")
-    val data = Data(project, project)
-    println(format.encodeToString(data))
-}        
+    // both Response and Project are abstract and their concrete subtypes are being serialized
+    val data: Response<Project> =  OkResponse(OwnedProject("kotlinx.serialization", "kotlin"))
+    val string = format.encodeToString(data)
+    println(string)
+    println(format.decodeFromString<Response<Project>>(string))
+}
+
