+++ guide/example/example-poly-17.kt
@@ -7,42 +7,24 @@ import kotlinx.serialization.json.*
 import kotlinx.serialization.modules.*
 
 @Serializable
-abstract class Response<out T>
-            
-@Serializable
-@SerialName("OkResponse")
-data class OkResponse<out T>(val data: T) : Response<T>()
-
-val responseModule = SerializersModule {
-    polymorphic(Response::class) {
-        subclass(OkResponse.serializer(PolymorphicSerializer(Any::class)))
-    }
-}
-
-val projectModule = SerializersModule {
-    fun PolymorphicModuleBuilder<Project>.registerProjectSubclasses() {
-        subclass(OwnedProject::class)
-    }
-    polymorphic(Any::class) { registerProjectSubclasses() }
-    polymorphic(Project::class) { registerProjectSubclasses() }
-}
-
-@Serializable
 abstract class Project {
     abstract val name: String
 }
-            
+
 @Serializable
 @SerialName("OwnedProject")
 data class OwnedProject(override val name: String, val owner: String) : Project()
 
-val format = Json { serializersModule = projectModule + responseModule }
-
-fun main() {
-    // both Response and Project are abstract and their concrete subtypes are being serialized
-    val data: Response<Project> =  OkResponse(OwnedProject("kotlinx.serialization", "kotlin"))
-    val string = format.encodeToString(data)
-    println(string)
-    println(format.decodeFromString<Response<Project>>(string))
+val module = SerializersModule {
+    polymorphic(Project::class) {
+        subclass(OwnedProject::class)
+    }
 }
 
+val format = Json { serializersModule = module }
+
+fun main() {
+    println(format.decodeFromString<Project>("""
+        {"type":"unknown","name":"example"}
+    """))
+}
