+++ guide/example/example-poly-18.kt
@@ -12,19 +12,26 @@ abstract class Project {
 }
 
 @Serializable
+data class BasicProject(override val name: String, val type: String): Project()
+
+@Serializable
 @SerialName("OwnedProject")
 data class OwnedProject(override val name: String, val owner: String) : Project()
 
 val module = SerializersModule {
     polymorphic(Project::class) {
         subclass(OwnedProject::class)
+        default { BasicProject.serializer() }
     }
 }
 
 val format = Json { serializersModule = module }
 
 fun main() {
-    println(format.decodeFromString<Project>("""
-        {"type":"unknown","name":"example"}
+    println(format.decodeFromString<List<Project>>("""
+        [
+            {"type":"unknown","name":"example"},
+            {"type":"OwnedProject","name":"kotlinx.serialization","owner":"kotlin"} 
+        ]
     """))
 }
