+++ guide/test/PolymorphismTest.kt
@@ -66,22 +66,22 @@ class PolymorphismTest {
     @Test
     fun testExamplePoly09() {
         captureOutput("ExamplePoly09") { example.examplePoly09.main() }.verifyOutputLinesStart(
-            "Exception in thread \"main\" kotlinx.serialization.SerializationException: Serializer for class 'Project' is not found.",
-            "Mark the class as @Serializable or provide the serializer explicitly."
+            "{\"type\":\"owned\",\"name\":\"kotlinx.coroutines\",\"owner\":\"kotlin\"}"
         )
     }
 
     @Test
     fun testExamplePoly10() {
         captureOutput("ExamplePoly10") { example.examplePoly10.main() }.verifyOutputLines(
-            "{\"type\":\"owned\",\"name\":\"kotlinx.coroutines\",\"owner\":\"kotlin\"}"
+            "{\"project\":{\"type\":\"owned\",\"name\":\"kotlinx.coroutines\",\"owner\":\"kotlin\"}}"
         )
     }
 
     @Test
     fun testExamplePoly11() {
-        captureOutput("ExamplePoly11") { example.examplePoly11.main() }.verifyOutputLines(
-            "{\"project\":{\"type\":\"owned\",\"name\":\"kotlinx.coroutines\",\"owner\":\"kotlin\"}}"
+        captureOutput("ExamplePoly11") { example.examplePoly11.main() }.verifyOutputLinesStart(
+            "Exception in thread \"main\" kotlinx.serialization.SerializationException: Serializer for class 'Any' is not found.",
+            "Mark the class as @Serializable or provide the serializer explicitly."
         )
     }
 
@@ -95,51 +95,43 @@ class PolymorphismTest {
 
     @Test
     fun testExamplePoly13() {
-        captureOutput("ExamplePoly13") { example.examplePoly13.main() }.verifyOutputLinesStart(
-            "Exception in thread \"main\" kotlinx.serialization.SerializationException: Serializer for class 'Any' is not found.",
-            "Mark the class as @Serializable or provide the serializer explicitly."
+        captureOutput("ExamplePoly13") { example.examplePoly13.main() }.verifyOutputLines(
+            "{\"type\":\"owned\",\"name\":\"kotlinx.coroutines\",\"owner\":\"kotlin\"}"
         )
     }
 
     @Test
     fun testExamplePoly14() {
         captureOutput("ExamplePoly14") { example.examplePoly14.main() }.verifyOutputLines(
-            "{\"type\":\"owned\",\"name\":\"kotlinx.coroutines\",\"owner\":\"kotlin\"}"
+            "{\"project\":{\"type\":\"owned\",\"name\":\"kotlinx.coroutines\",\"owner\":\"kotlin\"}}"
         )
     }
 
     @Test
     fun testExamplePoly15() {
         captureOutput("ExamplePoly15") { example.examplePoly15.main() }.verifyOutputLines(
-            "{\"project\":{\"type\":\"owned\",\"name\":\"kotlinx.coroutines\",\"owner\":\"kotlin\"}}"
+            "{\"project\":{\"type\":\"owned\",\"name\":\"kotlinx.coroutines\",\"owner\":\"kotlin\"},\"any\":{\"type\":\"owned\",\"name\":\"kotlinx.coroutines\",\"owner\":\"kotlin\"}}"
         )
     }
 
     @Test
     fun testExamplePoly16() {
         captureOutput("ExamplePoly16") { example.examplePoly16.main() }.verifyOutputLines(
-            "{\"project\":{\"type\":\"owned\",\"name\":\"kotlinx.coroutines\",\"owner\":\"kotlin\"},\"any\":{\"type\":\"owned\",\"name\":\"kotlinx.coroutines\",\"owner\":\"kotlin\"}}"
-        )
-    }
-
-    @Test
-    fun testExamplePoly17() {
-        captureOutput("ExamplePoly17") { example.examplePoly17.main() }.verifyOutputLines(
             "{\"type\":\"OkResponse\",\"data\":{\"type\":\"OwnedProject\",\"name\":\"kotlinx.serialization\",\"owner\":\"kotlin\"}}",
             "OkResponse(data=OwnedProject(name=kotlinx.serialization, owner=kotlin))"
         )
     }
 
     @Test
-    fun testExamplePoly18() {
-        captureOutput("ExamplePoly18") { example.examplePoly18.main() }.verifyOutputLinesStart(
+    fun testExamplePoly17() {
+        captureOutput("ExamplePoly17") { example.examplePoly17.main() }.verifyOutputLinesStart(
             "Exception in thread \"main\" kotlinx.serialization.json.internal.JsonDecodingException: Polymorphic serializer was not found for class discriminator 'unknown'"
         )
     }
 
     @Test
-    fun testExamplePoly19() {
-        captureOutput("ExamplePoly19") { example.examplePoly19.main() }.verifyOutputLines(
+    fun testExamplePoly18() {
+        captureOutput("ExamplePoly18") { example.examplePoly18.main() }.verifyOutputLines(
             "[BasicProject(name=example, type=unknown), OwnedProject(name=kotlinx.serialization, owner=kotlin)]"
         )
     }
